move new code to vendor.py
